pyproject.toml
[tool.poetry]
name = "guac-session"
version = "0.1.0"
description = ""
authors = ["enzok <7831008+enzok@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "guac_session"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
black = "^22.6.0"
Django = "^4.0.6"
django-extensions = "^3.2.0"
pyguacamole = "^0.11"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
Werkzeug = "^2.2.1"
gunicorn = "^20.1.0"
libvirt-python = "^8.6.0"
channels = "^3.0.5"
python-dotenv = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
